Fix type confusion in heap-numbers-equal? calls from VM
When heap-numbers-equal? is called from eqv?, we have already ensured
that the both objects have the same heap type.  However when called by
the VM, the precondition is just that both are heap numbers -- not
necessarily of the same type.  Fix to add an additional check in
heap-numbers-equal?.  Could cause crashers!

* libguile/eq.c (scm_i_heap_numbers_equal_p): Add additional check.