Accept ccache files as fresh if special file FINAL is present
When unpacking zip files or MSIX files, timestamps are not always
preserved.  When delivering a Guile instance as zip or MSIX,
when unpacked and first launched, Guile may try to recompile
all the system's *.go files since they do not necessarily have
timestamps greater than their associated *.scm files.

This commit adds the feature where if a file named FINAL is put
in %ccache-dir or %site-ccache-dir, the *.go files therein
are presumed fresh, and will not be recompiled when the scm files
are newer.

* libguile/load.c (compiled_is_final): new function
  (compiled_is_fresh): use new function to test for FINAL file