When FINAL file appears in SCM directory, don't recompile
When a file named FINAL exists in the ccache or site ccache
directory, modifying a scm
file in that directory or one of its subdirectories will not
cause a recompilation of the *.go files.  This is for the use case
of unpacking a distro from an archive that does not preserve
timestamps.

* libguile/load.c (compiled_is_final): new helper func
  (compiled_is_fresh): use new helper func