mpz->integer: do not convert to bignum if inum is sufficient
Before, `take_mpz` would convert to bignum if the `mpz` value
does not fit into a signed long, even though `scm_t_inum` is not
coupled to `long` anymore.  This leads to problems with functions
that expect values that numerically fit into a fixnum to be of
inum type, e.g., during bytecode compilation of long-long-immediates,
where `ash` returns a bignum, while the assembler requires an inum.

Call `scm_integer_from_mpz` instead, which will convert back to inum
if the number fits into an inum.

* libguile/integers.c (take_mpz): use `scm_integer_from_mpz`