00-socket.test: (flush-all-ports) before forking
Flush all the ports before forking, not just the current output and
error ports so that, for example, the parallel test harness' --trs-file
port will be included. Without that the 00-socket.trs test output ends
up with duplications.

Thanks to Dale P. Smith for noticing the problem and tracking down the
cause.

* test-suite/tests/00-socket.test (primitive-fork-if-available): flush
all ports before forking.