fluids.test: be more determined when checking fluid gc
The "fluids are GC'd" test failed on s390x. Investigation revealed that
adding a (sleep 0) allowed the test to succeed. Rework the test to try
more than once (up to 100 times), with a (sleep 0) between attempts.

The investigation also revealed that a (usleep n) for various values of
n, including 0 and 1000000 never worked. Further testing found that
changing usleep to convert the interval as a uint (as sleep does)
instead of a ulong allowed the test to succeed.

* test-suite/tests/fluids.test: Try up to 100 times, sleeping between
attempts.