Fix meta-command `describe' on macros
The `describe, meta-command will only lookup the current module for a
variable matching a name if the parsed form from the REPL is a symbol.

This works fine if the current reader is `read' but not with
`read-syntax'.

For example with read-syntax:

  $ echo ',describe current-filename' | guile
  [...]
  #f

But with read:

  $ echo '(fluid-set! current-reader read) ,describe current-filename' | guile
  [...]
  A macro that expands to the current filename: the filename that
  the (current-filename) form appears in.  Expands to #f if this
  information is unavailable.

To circumvent this, simply compare the form with `identifier?'.

* module/system/repl/command.scm (meta-command)[describe]: Check REPL
input against `identifier?' for syntaxes returned by `read-syntax'.